backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiofiles==23.2.1
python-multipart==0.0.9
pypdf==4.2.0
httpx==0.27.0
python-dotenv==1.0.1
